Prioritize location selection for fragments
When selecting a location for resolving Fields within Fragment Spreads,
we should select the parent's location if possible to avoid executing
federation unnecessarily. This fixes the problem where we used to simply
select the first element in the array. The proposed fix will prioritize
the parent's location. If no priority is matched, then the first
location is still selected. Note that this creates uneven traffic for
one location.